api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-docx==1.1.2
docxtpl==0.17.0
python-dateutil==2.9.0
pydantic==2.8.2
